requirements.txt
Django==1.11.29
django-timezone-field==3.1
djangorestframework==3.9.4
pkg-resources==0.0.0
pytz==2019.3
